feat: endpoint.defaults()
