+++ lib/get-defaults.js
@@ -1,8 +1,8 @@
 const getUserAgent = require('universal-user-agent')
 
-module.exports = defaults
+module.exports = getDefaults
 
-function defaults () {
+function getDefaults () {
   const version = require('../package.json').version
   const userAgent = `octokit-endpoint.js/${version} ${getUserAgent()}`
